+++ src/main/java/ru/yandex/practicum/filmorate/controllers/FilmController.java
@@ -1,16 +1,17 @@
 package ru.yandex.practicum.filmorate.controllers;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
+import ru.yandex.practicum.filmorate.validations.CheckValidFilm;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.IdGenerator;
 
 import javax.validation.Valid;
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
@@ -21,13 +22,18 @@ import java.util.Map;
 @RequestMapping("/films")
 public class FilmController {
 
-    IdGenerator id = new IdGenerator();
-    private final Map<Integer, Film> films = new HashMap<>();
-    private final LocalDate startFilmDate = LocalDate.of(1895, 12, 28);
+    private IdGenerator id = new IdGenerator();
+    private Map<Integer, Film> films = new HashMap<>();
+    private CheckValidFilm checkValidFilm;
+
+    @Autowired
+    public FilmController(CheckValidFilm checkValidFilm) {
+        this.checkValidFilm = checkValidFilm;
+    }
 
     @PostMapping
     public ResponseEntity<Film> create(@Valid @RequestBody Film film){
-        checkValidFilm(film, false);
+        checkValidFilm.checkValidFilm(film, false);
         log.info("Запрос получен к эндпоинту /films");
         film.setId(id.generator());
         if(films.containsKey(film.getId())){
@@ -57,11 +63,5 @@ public class FilmController {
         log.info("Запрос получен к эндпоинту /films");
         System.out.println("total films: " + films.size());
         return new ArrayList<>(films.values());
-    }
-
-    public void checkValidFilm(Film film, Boolean isCreated) {
-        if (film.getReleaseDate().isBefore(startFilmDate)) {
-            throw new ValidationException("Релиз раньше 28 декабря 1895 года");
-        }
     }
 }
