+++ src/main/java/ru/yandex/practicum/filmorate/controllers/UserController.java
@@ -2,14 +2,14 @@ package ru.yandex.practicum.filmorate.controllers;
 
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.IdGenerator;
+import ru.yandex.practicum.filmorate.validations.CheckValidUser;
 
 import javax.validation.Valid;
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
@@ -19,14 +19,20 @@ import java.util.Map;
 @RequestMapping("/users")
 
 public class UserController {
-    IdGenerator id = new IdGenerator();
 
-    Map<Long, User> users = new HashMap<>();
+    private IdGenerator id = new IdGenerator();
+    private Map<Long, User> users = new HashMap<>();
+    private CheckValidUser checkValidUser;
+
+    @Autowired
+    public UserController(CheckValidUser checkValidUser) {
+        this.checkValidUser = checkValidUser;
+    }
 
     @PostMapping
     public ResponseEntity<User> createUser(@Valid @RequestBody User user) {
         log.info("запрос получен к эндпоинту /users");
-        checkValidUser(user, true);
+        checkValidUser.checkValidUser(user, true);
         user.setId(id.generator());
         if (users.containsKey(user.getId())) {
             log.info("Ошибка добавления: " + user.getName());
@@ -52,20 +58,5 @@ public class UserController {
     @GetMapping
     public ArrayList<User> allUsers() {
         return new ArrayList<>(users.values());
-    }
-
-    private void checkValidUser(User user, Boolean isCreated){
-        if(user.getName()==null||user.getName().isBlank()||user.getName().isEmpty()){
-            user.setName(user.getLogin());
-        }
-        if(user.getLogin().isEmpty() || user.getLogin().isBlank()){
-            throw new ValidationException("Проверьте логин");
-        }
-        if(user.getBirthday().isAfter(LocalDate.now())){
-            throw new ValidationException("Проверьте дату рождения");
-        }
-        if(user.getEmail().isEmpty() || !user.getEmail().contains("@")){
-            throw new ValidationException("Проверьте почту");
-        }
     }
 }
